requirements.txt
APScheduler==3.11.0
bidict==0.23.1
blinker==1.9.0
boto3==1.35.82
botocore==1.35.82
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
curl_cffi==0.7.1
Deprecated==1.2.15
Flask==3.1.0
Flask-Limiter==3.9.2
# Flask-SocketIO==5.4.1
gevent==24.11.1
# gevent-websocket==0.10.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
limits==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mutagen==1.47.0
ordered-set==4.1.0
packaging==24.2
pycparser==2.22
pycryptodomex==3.21.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.0
# python-socketio==5.11.4
requests==2.32.3
rich==13.9.4
s3transfer==0.10.4
# simple-websocket==1.1.0
six==1.17.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
# websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
yt-dlp==2024.12.23
zope.event==5.0
zope.interface==7.2
